+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Changelog
 **Next release**
 
 * Toggle restriction will allow to choose between, clear, restrict and template ([issue](/../../issues/1319))
+* Option to select if existing restrictions should be deleted on fetch ([issue](/../../issues/1119))
 * Updated Dutch translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)